requirements.txt
torch
numpy
pandas
pillow
tqdm
